Final version for this project. Looks good.
Possible future add-ons (if someone ever wants to work on this well documented and with a great architecture project) would be to generate the GUI. To do so, the person could go through the XML of the client and create the buttons according to the defined commands. The problem with this is that even the cheat commands will be there. Then the person could create an Ant script to generate the GUI, after which it would be included into the client package. And then simply use the buildClient and buildServer ant scripts as usual.